backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sro-backend"
version = "0.1.0"
description = "Sovereign Robotics Ops - backend (FastAPI)"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "SQLAlchemy>=2.0.30",
  "httpx>=0.27.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.9"]
dev = ["pytest>=8.0.0", "ruff>=0.4.0"]
